+++ overlay/swbuilding.c
@@ -110,7 +110,7 @@ char swbuilding_cursor(swbuilding_t* w, game_t* g, int _x, int _y)
 	}
 	pos.y += 20;
 
-	for (int i = 0; i < t->n_items; i++)
+	for (size_t i = 0; i < t->n_items; i++)
 	{
 		transform_t* tr = &t->items[i];
 		if (tr->n_res == 0)
@@ -192,7 +192,7 @@ void swbuilding_draw(swbuilding_t* w, game_t* g)
 	pos.y += 20;
 
 	int cur_work = b->work_n == 0 ? -1 : b->work_list[0];
-	for (int i = 0; i < t->n_items; i++)
+	for (size_t i = 0; i < t->n_items; i++)
 	{
 		transform_t* tr = &t->items[i];
 		if (tr->n_res == 0)
@@ -206,7 +206,7 @@ void swbuilding_draw(swbuilding_t* w, game_t* g)
 
 		wchar_t* name = g->u->items[c->id].name;
 		wchar_t buffer[1024];
-		if (i == cur_work)
+		if ((int) i == cur_work)
 			swprintf(buffer, 1024, L"%ls (%i%%)", name, (int) floor(100*b->work_progress));
 		else
 			swprintf(buffer, 1024, L"%ls", name);
@@ -249,7 +249,7 @@ char swbuilding_catch(swbuilding_t* w, game_t* g, int _x, int _y, int t)
 	pos.y += 20; // building name
 	pos.y += 20; // material
 
-	for (int i = 0; i < b->t->n_items; i++)
+	for (size_t i = 0; i < b->t->n_items; i++)
 	{
 		transform_t* tr = &b->t->items[i];
 		if (tr->n_res == 0)
