+++ universe/item.c
@@ -26,15 +26,15 @@ void kindOf_item_init(kindOf_item_t* i, universe_t* u)
 	i->skill = 0;
 	i->category = 0;
 
-	for (int k = 0; k < N_SPECIAL_SKILLS; k++)
+	for (size_t k = 0; k < N_SPECIAL_SKILLS; k++)
 		i->bonus_special[k] = 0;
 
 	i->bonus_material = CALLOC(float, u->n_materials);
-	for (int k = 0; k < u->n_materials; k++)
+	for (size_t k = 0; k < u->n_materials; k++)
 		i->bonus_material[k] = 0;
 
 	i->bonus_item = CALLOC(float, u->n_iskills);
-	for (int k = 0; k < u->n_iskills; k++)
+	for (size_t k = 0; k < u->n_iskills; k++)
 		i->bonus_item[k] = 0;
 }
